requirements.txt
astroid==2.2.0
bootstrap-admin==0.4.0
Django==2.1.7
django-appconf==1.0.2
django-bootstrap3==9.1.0
django-select2==6.0.3
factory-boy==2.11.1
Faker==1.0.5
gunicorn==19.8.1
isort==4.3.9
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.2.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pylint==2.2.2
python-dateutil==2.6.1
pytz==2018.9
six==1.12.0
text-unidecode==1.2
typed-ast==1.2.0
whitenoise==3.3.1
wrapt==1.11.1
